requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Security-Too==5.3.1
Flask-Session==0.4.0
pymysql==1.1.0
gunicorn==21.2.0
